- refactor method
