+++ cloud/server/methods/user/subscribe_product.ts
@@ -1,54 +1,8 @@
-import {Users} from "../collections/Users";
-import {OM} from "../code/General/ObjectManager";
-import {User} from "../models/User";
-import {Role} from "../models/Role";
-import _ = require("lodash");
-import {UserHasLicense} from "../models/UserInterface";
-
-new ValidatedMethod({
-  name    : 'user.create_user_by_license',
-  validate: function (user) {
-    try {
-      new Users['$schema'].validate(user);
-    } catch (e) {
-      throw new Meteor.Error("Error", e.message);
-    }
-  },
-  run     : function (user) {
-    // Kiểm tra xem license đã có shop_owner_id chưa?
-    
-    // insert user with license
-    
-    return Users.insert(user);
-  }
-});
-DDPRateLimiter.addRule({
-                         userId: function (userId) {
-                           return true;
-                         },
-                         type  : "method",
-                         name  : "user.create_user_by_license",
-                       }, 1, 1000);
-
-new ValidatedMethod({
-  name    : 'user.get_roles',
-  validate: function () {
-    if (!this.userId) {
-      throw new Meteor.Error("Error", "Access denied");
-    }
-  },
-  run     : function () {
-    let userModel: User = OM.create<User>(User).loadById(this.userId);
-    return userModel.getRoles();
-  }
-});
-DDPRateLimiter.addRule({
-                         userId: function (userId) {
-                           return true;
-                         },
-                         type  : "method",
-                         name  : "user.get_roles",
-                       }, 1, 1000);
+import {User} from "../../models/User";
+import {OM} from "../../code/General/ObjectManager";
+import {Role} from "../../models/Role";
+import {UserHasLicense} from "../../models/UserInterface";
+import * as _ from "lodash";
 
 new ValidatedMethod({
   name    : "user.subscribe_product",
